% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{print.sz_block}
\alias{print.sz_block}
\title{Prints an `sz_block` object (any of the five block kinds).}
\usage{
\method{print}{sz_block}(x, ...)
}
\arguments{
\item{x}{An `sz_block` object (from [sz_float()], [sz_int()],
[sz_categorical()], [sz_binary()], or [sz_permutation()]).}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Prints an `sz_block` object (any of the five block kinds).
}
