% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{print.sz_result}
\alias{print.sz_result}
\title{Prints an `sz_result` object.}
\usage{
\method{print}{sz_result}(x, ...)
}
\arguments{
\item{x}{An `sz_result` object (from [.sz_wrap_result]/[Algorithm]'s
own `run()`).}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Prints an `sz_result` object.
}
