% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{print.sz_space}
\alias{print.sz_space}
\title{Prints an `sz_space` object.}
\usage{
\method{print}{sz_space}(x, ...)
}
\arguments{
\item{x}{An `sz_space` object (from [sz_space()]).}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Prints an `sz_space` object.
}
