% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{sezgi_version}
\alias{sezgi_version}
\title{Version of the underlying sezgi Rust core.}
\usage{
sezgi_version()
}
\value{
A character scalar with the crate version string.
}
\description{
Version of the underlying sezgi Rust core.
}
