% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{sz_stats_cliffs_delta}
\alias{sz_stats_cliffs_delta}
\title{Cliff's delta effect size for two independent (unpaired) samples.}
\usage{
sz_stats_cliffs_delta(a, b)
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector.}
}
\value{
A numeric scalar in `[-1, 1]`.
}
\description{
Cliff's delta effect size for two independent (unpaired) samples.
}
